+++ next.config.mjs
@@ -30,7 +30,6 @@ const commonHeaders = [
 ]
 
 const nextConfig = {
-  output: 'standalone',
   compress: true,
   poweredByHeader: false,
   async headers() {